added AJAX for adding and deleting Posts

+++ controllers/posts_controller.js
@@ -3,10 +3,18 @@ const Comment = require("../models/comment");
 
 module.exports.create = async function (req, res) {
   try {
-    await Post.create({
+    let post = await Post.create({
       content: req.body.content,
       user: req.user._id,
     });
+    if (req.xhr) {
+      return res.status(200).json({
+        data: {
+          post,
+        },
+        message: "Post Published!",
+      });
+    }
     req.flash("success", "Post Published!");
     return res.redirect("back");
   } catch (err) {
@@ -21,6 +29,15 @@ module.exports.destroy = async function (req, res) {
     // .id means converting the object _id into string
     if (post.user == req.user.id) {
       post.remove();
+
+      if (req.xhr) {
+        return res.status(200).json({
+          data: {
+            post_id: req.params.id,
+          },
+          message: "Post Deleted",
+        });
+      }
       await Comment.deleteMany({ post: req.params.id });
       req.flash("success", "Post Deleted!");
       return res.redirect("back");
@@ -29,7 +46,7 @@ module.exports.destroy = async function (req, res) {
       return res.redirect("back");
     }
   } catch (err) {
-    req.flash("error", err);
+    req.flash("error", err.message);
     return res.redirect("back");
   }
 };
